added sortPaginate middleware
