+++ controllers/articleController.js
@@ -5,69 +5,7 @@ import User from '../models/user.js';
 
 const getArticles = async (req, res, next) => {
   try {
-    let query;
-
-    // Make a copy of req query in order to select only certain fields
-    const reqQuery = { ...req.query };
-
-    // Fields to esclude
-    const fieldsToRemove = ['select', 'sort', 'page', 'limit'];
-
-    // loop over fields to exclude and remove them from reqQuery
-    fieldsToRemove.forEach(param => delete reqQuery[param]);
-
-    // Create express friendly query strings, that will prepend a '$' to the operators
-    const queryStr = JSON.stringify(reqQuery).replace(
-      /\b(gt|gte|lt|lte|in)\b/g,
-      match => `$${match}`
-    );
-
-    // Finding resource
-    query = Article.find(JSON.parse(queryStr));
-
-    // Select fields
-    if (req.query.select) {
-      const fields = req.query.select.split(',').join(' ');
-      query = query.select(fields);
-    }
-
-    // Sort
-    if (req.query.sort) {
-      const sortBy = req.query.sort.split(',').join(' ');
-      query = query.sort(sortBy);
-    } else {
-      query.sort('-createdAt');
-    }
-
-    // Pagination
-    const page = parseInt(req.query.page, 10) || 1;
-    const limit = parseInt(req.query.limit, 10) || 3;
-    const startIndex = (page - 1) * limit;
-    const endIndex = page * limit;
-    const total = await Article.countDocuments();
-
-    query = query.skip(startIndex).limit(limit);
-
-    // Execute query
-    const articles = await query;
-
-    // Pagination results
-    const pagination = {};
-
-    if (endIndex < total) {
-      pagination.next = {
-        page: page + 1,
-        limit,
-      };
-    }
-
-    if (startIndex > 0) {
-      pagination.prev = { page: page - 1, limit };
-    }
-
-    res
-      .status(200)
-      .json({ length: articles.length, pagination, data: articles });
+    res.status(200).json(res.sortPaginate);
   } catch (err) {
     next(err);
   }
